fix VPP v2302 build
- include header files
- update DPDK plugin by using DPDK latest API's
- fix ‘struct rte_eth_conf’ has no member named ‘fdir_conf’
- fix data type, updated from u16 to u32 of the variable
- fix format dpdk rte_device
- fix other build errors according to update in dpdk 22.11

DPDK-3834

+++ src/vlib/pci/pci.h
@@ -48,7 +48,7 @@ typedef CLIB_PACKED (union
 {
   struct
     {
-      u16 domain;
+      u32 domain;
       u8 bus;
       u8 slot: 5;
       u8 function:3;
